+++ docs/pages/_en/1.0/reference/release-notes.md
@@ -20,6 +20,7 @@ group: Deployment & Configs
 - Added 'Category Graphic Overrides' component.
 - Added 'Element Graphic Overrides' component.
 - Added 'Filter Graphic Overrides' component.
+- Added 'View Filters' component.
 
 {% endcapture %}
 {% include ltr/release_header_next.html title="Upcoming Changes" note=rc_release_notes %}